Use Plural Strings Instead of Normal Strings for ForgotPw Fragments Timer Error

+++ app/src/main/java/com/xeniac/warrantyroster_manager/ui/landing/fragments/ForgotPwFragment.kt
@@ -177,7 +177,7 @@ class ForgotPwFragment : Fragment(R.layout.fragment_forgot_pw) {
                                     ) { snackbar?.dismiss() }
                                 }
                                 it.contains(ERROR_TIMER_IS_NOT_ZERO) -> {
-                                    val seconds = viewModel.timerInMillis / 1000
+                                    val seconds = (viewModel.timerInMillis / 1000).toInt()
                                     snackbar = showTimerIsNotZeroError(
                                         requireContext(),
                                         requireView(),
